chore(storybook): organize icon loader into tooling
