+++ .storybook/decorators/icon-sprites.js
@@ -1,0 +1,36 @@
+// Used in the icon sprite decorator to inject the sprite sheets into the document
+import workflowSprite from "@adobe/spectrum-css-workflow-icons/dist/spectrum-icons.svg?raw";
+import uiSprite from "@spectrum-css/ui-icons/dist/spectrum-css-icons.svg?raw";
+import { makeDecorator, useEffect } from "@storybook/preview-api";
+
+/**
+ * @type import('@storybook/csf').DecoratorFunction<import('@storybook/web-components').WebComponentsFramework>
+ **/
+export const withIconSpriteSheet = makeDecorator({
+	name: "withIconSpriteSheet",
+	parameterName: "spritesheet",
+	wrapper: (StoryFn, context) => {
+		const {
+            loaded = {},
+		} = context;
+
+        // Load the icons into the window object
+        if (loaded.icons) window.icons = loaded.icons;
+
+		useEffect(() => {
+            // Inject the sprite sheets into the document
+            let sprite = document.getElementById("spritesheets");
+            if (!sprite) {
+                sprite = document.createElement("div");
+                sprite.id = "spritesheets";
+                sprite.innerHTML = workflowSprite + uiSprite;
+                document.body.appendChild(sprite);
+            }
+            else {
+                sprite.innerHTML = workflowSprite + uiSprite;
+            }
+		}, []);
+
+        return StoryFn(context);
+	},
+});
